requirements.txt
paho-mqtt
SpeechRecognition
pygame
edge-tts
vosk
ollama
pyautogui
screen-brightness-control
